Reverted System.DoubleNumerics change

+++ Core/Core/Serialisation/BaseObjectSerializerV2.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.DoubleNumerics;
+using System.Numerics;
 using System.Reflection;
 using System.Text.RegularExpressions;
 using System.Threading;
@@ -158,7 +158,7 @@ namespace Speckle.Core.Serialisation
       }
       if (obj is Matrix4x4 m)
       {
-        return new List<double>()
+        return new List<float>()
         {
           m.M11, m.M12, m.M13, m.M14,
           m.M21, m.M22, m.M23, m.M24,
